issue-121 extract Adapter.fix spec into a separate file
