+++ src/component/interfaces/index.ts
@@ -1,5 +1,5 @@
 import { Datasource, DatasourceGet } from './datasource';
-import { ItemAdapter, ItemsPredicate, ItemsIteratorCallBack, ClipOptions, FixOptions, Adapter } from './adapter';
+import { ItemAdapter, ItemsPredicate, ItemsLooper, ClipOptions, FixOptions, Adapter } from './adapter';
 import { Settings, DevSettings } from './settings';
 import { Direction } from './direction';
 import { WindowScrollState, ScrollEventData, ScrollState, SyntheticScroll, WorkflowOptions, State } from './state';
@@ -10,7 +10,7 @@ export {
   DatasourceGet,
   ItemAdapter,
   ItemsPredicate,
-  ItemsIteratorCallBack,
+  ItemsLooper,
   ClipOptions,
   FixOptions,
   Adapter,
